Updated player assets and equip function
Instead of saving player's gear and inventory as separate files, class PlayerAssets has been created so that there will not be a need for many files. PlayerAssets contains player's gold amount, gear and inventory. Might add more things inside PlayerAssets. 

Equip function has also been updated to work with the newly created PlayerAssets and is now part of the class as a function.